vector3 class
