--- BMC-ENGINE/classes/math/Vector2.cpp
@@ -1,13 +1,0 @@
-#include "Vector2.h"
-
-Vector2::Vector2(int x, int y)
-{
-	this->x = x;
-	this->y = y;
-}
-
-Vector2::Vector2()
-{
-	this->x = 0;
-	this->y = 0;
-}
