--- BMC-ENGINE/classes/math/Vector2.h
@@ -1,13 +1,0 @@
-#ifndef VECTOR2_H
-#define VECTOR2_H
-
-class Vector2 {
-public:
-	int x;
-	int y;
-	
-	Vector2(int x, int y);
-	Vector2();
-};
-
-#endif // VECTOR2_H
